src/presentation/middlewares/auth/auth-middleware.ts
import {
  AccessDeniedError,
  forbidden,
  HttpRequest,
  HttpResponse,
  LoadAccountByToken,
  Middleware,
  ok, serverError
} from './auth-middleware-protocols'

export class AuthMiddleware implements Middleware {
  constructor (
    private readonly loadAccountByToken: LoadAccountByToken
  ) {}

  async handle (httpRequest: HttpRequest): Promise<HttpResponse> {
    try {
      const accessToken = httpRequest.headers?.['x-access-token']

      if (accessToken) {
        const account = await this.loadAccountByToken.load(accessToken)
        if (account) {
          return ok({ accountId: account.id })
        }
      }

      return forbidden(new AccessDeniedError())
    } catch (e) {
      return serverError(e)
    }
  }
}
